Correct preprototype template
This is a corrected version of the original template. I have changed the player's collider shape from a circle to a box, and I also shortened the width of the ground check boxcast so that it doesn't perceive the side of platforms as ground (this is partially what caused the climbing bug, the other half of the bug is related to friction).